adding favorites to user[bugs], change the state

+++ backend/routes/comment.js
@@ -3,19 +3,6 @@ const Post = require("../models/post.model");
 const Comment = require("../models/comment.model");
 const isLoggedIn  = require("../middlewares/middleware");
 
-
-
-//Get comments
-router.get("/:postId", (req,res) => {
-    Post.findById(req.params.postId)
-        .then(post => {
-            res.json(post.comments)
-        })
-        .catch(err => {
-            res.status(400).json('Cannot post yout comment')
-        })
-});
-
 //Add comment
 router.post("/:postId", isLoggedIn, async (req, res) => {
     try{
@@ -37,15 +24,26 @@ router.post("/:postId", isLoggedIn, async (req, res) => {
     }
 });
 
-
-router.delete("/:commentId",  (req, res) => {
-    Comment.findByIdAndRemove(req.params.commentId)
-        .then(comment => {
+//Remove Comment
+router.delete("/:postId/:commentId", isLoggedIn, async (req, res) => {
+    try {
+        const comment = await Comment.findById(req.params.commentId);
+        const post = await Post.findById(req.params.postId)
+        //find if the author of the comment is the current user
+        if(comment.author.equals(req.user)){
+            //remove spacific comment from the post object
+            const newComments = await post.comments.filter(newComment => !newComment.equals(req.params.commentId))
+            post.comments = newComments
+            await post.save();
+            //delete the comment from the database
+            await Comment.findByIdAndRemove(req.params.commentId)
             res.json('Comment removed successfully!')
-        })
-        .catch(err => {
-            res.status(400).json('cannot delete your comment')
-        })
+        }else{
+            res.status(400).json("Error: " + err)
+          }
+    }catch(err){
+        res.status(400).json('cannot delete your comment')
+    }
 });
